+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Release Notes
 
 Version 6.3.5
 -------------
+ * Clear database and re-sync play information
+ * Syncing improvements
  * Bug fixes
 
 Version 6.3.4